Winter code cleanup!

+++ java/iceandshadow2/nyx/blocks/ore/NyxBlockOreDevora.java
@@ -52,19 +52,19 @@ public class NyxBlockOreDevora extends NyxBlockOre {
 			int par3, int par4, Random par5Random) {
 		if(par5Random.nextBoolean())
 			return;
-		double var9 = (double)((float)par3 + par5Random.nextFloat());
+		double var9 = par3 + par5Random.nextFloat();
 		double var13 = 0.0D;
 		double var15 = 0.0D;
 		double var17 = 0.0D;
 		int var19 = par5Random.nextInt(2) * 2 - 1;
 		int var20 = par5Random.nextInt(2) * 2 - 1;
-		var13 = ((double)par5Random.nextFloat() - 0.5D) * 0.125D;
-		var15 = ((double)par5Random.nextFloat() - 0.5D) * 0.125D;
-		var17 = ((double)par5Random.nextFloat() - 0.5D) * 0.125D;
-		double var11 = (double)par4 + 0.5D + 0.25D * (double)var20;
-		var17 = (double)(par5Random.nextFloat() * 1.0F * (float)var20);
-		double var7 = (double)par2 + 0.5D + 0.25D * (double)var19;
-		var13 = (double)(par5Random.nextFloat() * 1.0F * (float)var19);
+		var13 = (par5Random.nextFloat() - 0.5D) * 0.125D;
+		var15 = (par5Random.nextFloat() - 0.5D) * 0.125D;
+		var17 = (par5Random.nextFloat() - 0.5D) * 0.125D;
+		double var11 = par4 + 0.5D + 0.25D * var20;
+		var17 = par5Random.nextFloat() * 1.0F * var20;
+		double var7 = par2 + 0.5D + 0.25D * var19;
+		var13 = par5Random.nextFloat() * 1.0F * var19;
 		par1World.spawnParticle("lava", var7, var9, var11, var13, var15, var17);
 	}
